requirements.txt
django-bootstrap-datepicker==1.2.3
django-google-tag-manager==0.0.3
django-multi-form-view==2.0.0
django-translation-manager==0.5.1
Django==1.11.2
PyJWT==1.6.1
argparse==1.2.1
certifi==2018.1.18
chardet==3.0.4
django-bootstrap-datepicker==1.2.3
django-google-tag-manager==0.0.3
django-multi-form-view==2.0.0
django-translation-manager==0.5.1
django-static-jquery==2.1.4
idna==2.6
mysqlclient==1.3.10
oauthlib==2.0.7
pip==9.0.1
polib==1.0.8
python-openid==2.2.5
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
setuptools==36.0.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.22
wheel==0.29.0
